setDisplayValueFromModel() should be abstract in BaseStringWidget and other classes should override it
