+++ collect_app/src/main/java/org/odk/collect/android/widgets/DecimalWidget.java
@@ -58,7 +58,22 @@ public class DecimalWidget extends StringWidget {
             fa[0] = new InputFilter.LengthFilter(19);
         }
         answerText.setFilters(fa);
+    }
 
+    private Double getDoubleAnswerValue() {
+        IAnswerData dataHolder = getFormEntryPrompt().getAnswerValue();
+        Double d = null;
+        if (dataHolder != null) {
+            Object dataValue = dataHolder.getValue();
+            if (dataValue != null) {
+                d = dataValue instanceof Integer ? (double) (Integer) dataValue : (Double) dataValue;
+            }
+        }
+        return d;
+    }
+
+    @Override
+    public void setDisplayValueFromModel() {
         Double d = getDoubleAnswerValue();
 
         if (d != null) {
@@ -72,18 +87,6 @@ public class DecimalWidget extends StringWidget {
             answerText.setText(nf.format(d));
             Selection.setSelection(answerText.getText(), answerText.getText().length());
         }
-    }
-
-    private Double getDoubleAnswerValue() {
-        IAnswerData dataHolder = getFormEntryPrompt().getAnswerValue();
-        Double d = null;
-        if (dataHolder != null) {
-            Object dataValue = dataHolder.getValue();
-            if (dataValue != null) {
-                d = dataValue instanceof Integer ? (double) (Integer) dataValue : (Double) dataValue;
-            }
-        }
-        return d;
     }
 
     @NonNull
