+++ collect_app/src/main/java/org/odk/collect/android/widgets/ExDecimalWidget.java
@@ -58,7 +58,22 @@ public class ExDecimalWidget extends ExStringWidget {
         InputFilter[] fa = new InputFilter[1];
         fa[0] = new InputFilter.LengthFilter(15);
         answerText.setFilters(fa);
+    }
 
+    private Double getDoubleAnswerValue() {
+        IAnswerData dataHolder = getFormEntryPrompt().getAnswerValue();
+        Double d = null;
+        if (dataHolder != null) {
+            Object dataValue = dataHolder.getValue();
+            if (dataValue != null) {
+                d = dataValue instanceof Integer ? (double) (Integer) dataValue : (Double) dataValue;
+            }
+        }
+        return d;
+    }
+
+    @Override
+    public void setDisplayValueFromModel() {
         Double d = getDoubleAnswerValue();
 
         if (d != null) {
@@ -73,18 +88,6 @@ public class ExDecimalWidget extends ExStringWidget {
 
             Selection.setSelection(answerText.getText(), answerText.getText().length());
         }
-    }
-
-    private Double getDoubleAnswerValue() {
-        IAnswerData dataHolder = getFormEntryPrompt().getAnswerValue();
-        Double d = null;
-        if (dataHolder != null) {
-            Object dataValue = dataHolder.getValue();
-            if (dataValue != null) {
-                d = dataValue instanceof Integer ? (double) (Integer) dataValue : (Double) dataValue;
-            }
-        }
-        return d;
     }
 
     @Override
