+++ collect_app/src/main/java/org/odk/collect/android/widgets/StringWidget.java
@@ -17,6 +17,7 @@ package org.odk.collect.android.widgets;
 import android.annotation.SuppressLint;
 import android.content.Context;
 
+import android.text.Selection;
 import android.text.method.TextKeyListener;
 import android.text.method.TextKeyListener.Capitalize;
 import android.view.Gravity;
@@ -56,6 +57,16 @@ public class StringWidget extends BaseStringWidget {
         }
 
         addAnswerView(answerText);
+    }
+
+    @Override
+    public void setDisplayValueFromModel() {
+        String currentAnswer = getFormEntryPrompt().getAnswerText();
+
+        if (currentAnswer != null) {
+            answerText.setText(currentAnswer);
+            Selection.setSelection(answerText.getText(), answerText.getText().toString().length());
+        }
     }
 
     @Override
